// frontend/src/sections/master-lists/masterListLines.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type MasterListLinesVariables = {
  storeId: string;
  masterListId: string;
  sort?: Array<{
    key: "name" | "code";
    desc?: boolean | null;
  }> | null;
  page?: {
    first?: number | null;
    offset?: number | null;
  } | null;
};

export type MasterListLinesResult = {
  masterListLines: ({
  __typename: "MasterListLineConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  item: {
  id: string;
  code: string;
  name: string;
  unitName: string | null;
};
}>;
});
};

export const MasterListLines = {
  query: "query masterListLines($storeId: String!, $masterListId: String!, $sort: [MasterListLineSortInput!], $page: PaginationInput) {\n  masterListLines(\n    storeId: $storeId\n    masterListId: $masterListId\n    sort: $sort\n    page: $page\n  ) {\n    ... on MasterListLineConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        item {\n          id\n          code\n          name\n          unitName\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<MasterListLinesResult, MasterListLinesVariables>;